Admin Console für Präsentation
-Nur 2 Boxen
-Bugfixes
-Bei Box.Create letzte Sequence holen

+++ 00_llbox/06_WiringPi/PIButton.cpp
@@ -76,7 +76,7 @@ void isr_General(int index, int dir)
 
 			pibuttons[index]->cnt = pibuttons[index]->cnt + dir; //dir = -1 oder +1			
 
-			if (abs(pibuttons[index]->cnt) >= pibuttons[index]->deltaCnt)
+			if ((unsigned int)abs(pibuttons[index]->cnt) >= pibuttons[index]->deltaCnt)
 			{
 				if (pibuttons[index]->ButtonEvent != 0)
 				{
